Moved subject_extensions to the Subject
