+++ lib/expectations_hook.rb
@@ -3,7 +3,6 @@ require 'mumukit/inspection'
 require 'yaml'
 
 require_relative 'stones_spec'
-require_relative 'subject_extensions'
 require_relative 'with_test_parser'
 
 module EvalExpectationsOnAST
